requirements.txt
streamlit
google-generativeai
google-cloud-bigquery
pandas
db-dtypes
python-dotenv
plotly